Implement base design class & remove design interface
